Reorganize component hierarchy into sub-directories
