+++ client/src/components/content/Stars.jsx
