+++ client/src/components/content/TitleView.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Link from './Link.jsx';
+import Link from '../styledComponents/Link.jsx';
 
 const Title = styled.div`
   h1 {
