+++ client/src/components/content/ratingsGraphic/RatingsBar.jsx
