+++ client/src/components/content/ratingsGraphic/RatingsGraphic.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import RatingsRow from './RatingsRow.jsx';
-import Link from './Link.jsx';
+import Link from '../../styledComponents/Link.jsx';
 
 const weighted = (one, two, three, four, five, total) => {
   let weightedAvg = (one * 1 + two * 2 + three * 3 + four * 4 + five * 5) / (total * 5);
