+++ client/src/components/picture/PictureDisplay.jsx
