+++ client/src/components/picture/PictureIcon.jsx
