+++ client/src/components/picture/PicturesView.jsx
